% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.r
\name{envi_checkpoint}
\alias{envi_checkpoint}
\title{Checkpoint the Current Package Configuration}
\usage{
envi_checkpoint(handle = envi_current_handle())
}
\arguments{
\item{handle}{the handle to the envronment you'd like to checkpoint.}
}
\description{
Checkpoint the Current Package Configuration
}
