% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.r
\name{envi_clone}
\alias{envi_clone}
\title{Clone an R Environment}
\usage{
envi_clone(path, handle = basename(path), verbose = TRUE, progress = verbose)
}
\arguments{
\item{path}{the path of the repository housing the R environment to clone.}

\item{handle}{the handle for the new environment.}

\item{verbose}{should extra information be printed? (Default TRUE)}

\item{progress}{should the progress of the clone be shown? (Default verbose)}
}
\description{
Clone an R Environment
}
