% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.r
\name{envi_create}
\alias{envi_create}
\title{Create an R Environment}
\usage{
envi_create(handle, full_name = handle, bare = FALSE, git_init = TRUE)
}
\arguments{
\item{handle}{the name of the new environment.}

\item{full_name}{the name of the environment directory. (Defalt is the
value of the handle argument)}

\item{bare}{should the project be initialized without attempting to 
discover and install R package dependencies? (Default TRUE)}

\item{git_init}{project include an initialized git repository? 
(Default TRUE)}
}
\description{
Create an R Environment
}
