% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.r
\name{envi_deactivate}
\alias{envi_deactivate}
\title{Deactivate an Environment}
\usage{
envi_deactivate(snapshot = TRUE, confirm = interactive(), force = TRUE)
}
\arguments{
\item{snapshot}{should a snapshot be created on exit? (Default TRUE)}

\item{confirm}{should the user be prompted before taking action? 
(Default interactive()).}

\item{force}{should the lockfile be generated even when preflight validation
check have failed? (Default FALSE)}
}
\description{
Deactivate an Environment
}
