% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-info.r
\name{envi_env_info}
\alias{envi_env_info}
\title{Get an Environments Packages and Versions}
\usage{
envi_env_info(handle)
}
\arguments{
\item{handle}{the handle of the environment to get the package list for.}
}
\description{
Get an Environments Packages and Versions
}
