% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.r
\name{envi_env_path}
\alias{envi_env_path}
\title{Get the path to an Environment}
\usage{
envi_env_path(handle = envi_current_handle())
}
\arguments{
\item{handle}{the handle of the environment. If missing, all paths are 
returned. Default is the current, activated environment handle, or NULL
if one is not activated.}
}
\description{
Get the path to an Environment
}
