% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.r
\name{envi_hard_reset}
\alias{envi_hard_reset}
\title{Reset a Remote Environment}
\usage{
envi_hard_reset(handle, clean = FALSE)
}
\arguments{
\item{handle}{the environment handle.}

\item{clean}{should file that are not part of the repository be removed?
(Default FALSE)}
}
\description{
Reset a Remote Environment
}
