% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.r
\name{envi_install_local_compressed}
\alias{envi_install_local_compressed}
\title{Install a Local Compressed R Environment}
\usage{
envi_install_local_compressed(
  path,
  handle = basename(path),
  verbose = TRUE,
  progress = verbose
)
}
\arguments{
\item{path}{the path of a compressed environment.}

\item{handle}{the handle for the new environment.}

\item{verbose}{should extra information be printed? (Default TRUE)}

\item{progress}{should the progress of the clone be shown? (Default verbose)}
}
\description{
Install a Local Compressed R Environment
}
