% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-info.r
\name{envi_list}
\alias{envi_list}
\title{List the Current Available Environments}
\usage{
envi_list(env_info = FALSE)
}
\arguments{
\item{env_info}{should individual environment information be included? 
(Default FALSE)}
}
\description{
List the Current Available Environments
}
