% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{envi_log}
\alias{envi_log}
\title{Log an Environment's Packages}
\usage{
envi_log(x, handle = envi_current_handle())
}
\arguments{
\item{x}{a envi log object.}

\item{handle}{the environment handle.}
}
\description{
Log an Environment's Packages
}
