% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.r
\name{envi_pb_install}
\alias{envi_pb_install}
\title{Install a Piggyback'ed Environment}
\usage{
envi_pb_install(
  file,
  repo,
  handle = basename(file),
  tag = "latest",
  verbose = TRUE,
  progress = verbose,
  keep = FALSE
)
}
\arguments{
\item{file}{name or vector of names of files to be downloaded.}

\item{repo}{Repository name in format "owner/repo". Will guess the
current repo if not specified.}

\item{handle}{the handle for the new environment.}

\item{tag}{tag for the GitHub release to which this data is attached.
(Default "latest")}

\item{verbose}{should extra information be printed? (Default TRUE)}

\item{progress}{should the progress of the clone be shown? (Default verbose)}

\item{keep}{should the original downloaded environment be kept? 
(Default FALSE)}
}
\description{
Install a Piggyback'ed Environment
}
