% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{envi_save_log}
\alias{envi_save_log}
\title{Save the Package Log}
\usage{
envi_save_log(
  pkg_log,
  location = file.path(get_envi_path(), "package-log.rds")
)
}
\arguments{
\item{pkg_log}{the package log.}

\item{location}{the location where the log should be saved. 
(Default ~/.envi/package-log.rds)}
}
\description{
Save the Package Log
}
