% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.r
\name{envi_uninstall}
\alias{envi_uninstall}
\title{Uninstall an Environment}
\usage{
envi_uninstall(handle, confirm = interactive(), purge = TRUE)
}
\arguments{
\item{handle}{the environment handle.}

\item{confirm}{should the user be prompted before removing the environment?
(Default TRUE)}

\item{purge}{should all files in the environments directory be removed?
(Default TRUE)}
}
\description{
Uninstall an Environment
}
