% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.r
\name{envi_url_install}
\alias{envi_url_install}
\title{Install a Remote Environment}
\usage{
envi_url_install(
  path,
  handle = basename(path),
  verbose = TRUE,
  progress = verbose
)
}
\arguments{
\item{path}{the path of the repository housing the R environment to clone.}

\item{handle}{the handle for the new environment.}

\item{verbose}{should extra information be printed? (Default TRUE)}

\item{progress}{should the progress of the clone be shown? (Default verbose)}
}
\description{
Install a Remote Environment
}
