% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlrren.R
\name{perlrren}
\alias{perlrren}
\title{Spatially perturb an ecological niche model that uses a log relative risk surface}
\usage{
perlrren(
  obs_ppp,
  covariates,
  predict = TRUE,
  predict_locs = NULL,
  radii = NULL,
  n_sim = 2,
  alpha = 0.05,
  parallel = FALSE,
  n_core = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{obs_ppp}{Input object of class 'ppp' a marked point pattern of presence and absence observations with 5 (five) features (columns): 1) ID, 2) longitude, 3) latitude, 4) presence/absence binary variable, 5) ordinal ID for spatial perturbation.}

\item{covariates}{Input object of class 'imlist' of 2 (two) covariates within the same spatial window and in the same coordinate reference system as \code{obs_ppp}.}

\item{predict}{Logical. If TRUE (the default), will predict the ecological niche in geographic space. If FALSE, will not predict.}

\item{predict_locs}{Input data frame of prediction locations with 4 features (columns): 1) longitude, 2) latitude, 3) covariate 1 as x-coordinate, 4) covariate 2 as y-coordinate. If unspecified (the default), automatically computed from an 'im' object within \code{covariates}.}

\item{radii}{Vector of length equal to the number of levels of ordinal ID in \code{obs_ppp}. Specifies the radii of the spatial perturbation at each level in units equivalent to the coordinate reference system of \code{obs_ppp}.}

\item{n_sim}{Integer, specifying the number of simulation iterations to perform.}

\item{alpha}{Numeric. The two-tailed alpha level for significance threshold (default is 0.05).}

\item{parallel}{Logical. If TRUE, will execute the function in parallel. If FALSE (the default), will not execute the function in parallel.}

\item{n_core}{Optional. Integer specifying the number of CPU cores on current host to use for parallelization (the default is 2 cores).}

\item{verbose}{Logical. If TRUE (the default), will print function progress during execution. If FALSE, will not print.}

\item{...}{Arguments passed to \code{\link{lrren}}.}
}
\value{
An object of class "list". This is a named list with the following components:

\describe{
\item{\code{sim}}{An object of class 'list' for the summary statistics of the iterative ecological niche.}
\item{\code{predict}}{An object of class 'ppp' a marked point pattern with summary statistics for the iterative ecological niche in geographic space.}
}

The returned \code{sim} is a named list with the following components:

\describe{
\item{\code{lrr_mean}}{An object of class 'im' for the mean log relative risk surface.}
\item{\code{lrr_sd}}{An object of class 'im' for the standard deviation of log relative risk surface.}
\item{\code{pval_mean}}{An object of class 'im' for the mean p-value surface.}
\item{\code{pval_prop}}{An object of class 'im' for the proportion of iterations were statistically significant surface.}
}

If \code{predict = FALSE}, the returned \code{predict} is empty. If \code{predict = TRUE}, the returned \code{predict} is an object of class 'ppp' a marked point pattern with the following features:

\describe{
\item{\code{x}}{Values for x-coordinate in geographic space (e.g., longitude).}
\item{\code{y}}{Values for y-coordinate in geographic space (e.g., latitude).}
\item{\code{v}}{Values for x-coordinate in covariate space.}
\item{\code{z}}{Values for x-coordinate in covariate space.}
\item{\code{lrr_mean}}{Values for the mean log relative risk surface.}
\item{\code{lrr_sd}}{Values for the standard deviation of log relative risk surface.}
\item{\code{pval_mean}}{Values for the mean p-value surface.}
\item{\code{pval_prop}}{Values for the proportion of iterations were statistically significant surface.}
}
}
\description{
Estimates the ecological niche of a single species with presence/absence data and two covariates, iteratively, by randomly perturbing ('jittering') the coordinates of observations.
}
\details{
This function performs a sensitivity analysis of an ecological niche model of a single species (presence/absence data), or the presence of one species relative to another, that uses two covariates. The observation locations (presence and absence data) are randomly spatially perturbed (i.e., "jittered") uniformly within a circular disc of a specified radius centered at their recorded location using the \code{\link[spatstat]{rjitter}} function. This method simulates the spatial uncertainty of observations, how that may affect the covariate values at each observation (i.e., misclassification error), and the estimated ecological niche based on the two specified covariates. Observations can be grouped into categories of uncertainty of class 'factor' and can vary by degrees of uncertainty specified using the \code{radii} argument.

The function iteratively estimates the ecological niche using the \code{\link{lrren}} function and computes four summary statistics at every grid cell (i.e., knot) of the estimated surface: 1) mean of the log relative risk, 2) standard deviation of the log relative risk, 3) mean of the asymptotically normal p-value, and 4) proportion of iterations were statistically significant based on a two-tailed alpha-level threshold (argument \code{alpha}). The process can be performed in parallel if \code{parallel = TRUE} using the \code{\link[foreach]{foreach}} function. The computed surfaces can be visualized using the \code{\link{plot_perturb}} function. If \code{predict = TRUE}, this function will predict the four summary statistics at every location specified with \code{predict_locs} and can also be visualized using the \code{\link{plot_perturb}} function.

For more information about the spatial perturbation, please refer to the \code{\link[spatstat]{rjitter}} function documentation.
}
\examples{
if (interactive()) {
  set.seed(1234) # for reproducibility

# Using the 'bei' and 'bei.extra' data within {spatstat.data}

# Covariate data (centered and scaled)
  ims <- spatstat.data::bei.extra
  ims[[1]]$v <- scale(ims[[1]]$v)
  ims[[2]]$v <- scale(ims[[2]]$v)
  
# Presence data
  presence <- spatstat.data::bei
  spatstat::marks(presence) <- data.frame("presence" = rep(1, presence$n),
                                              "lon" = presence$x,
                                              "lat" = presence$y)
                                          
# (Pseudo-)Absence data
  absence <- spatstat::rpoispp(0.008, win = ims[[1]])
  spatstat::marks(absence) <- data.frame("presence" = rep(0, absence$n),
                                              "lon" = absence$x,
                                              "lat" = absence$y)
# Combine into readable format
  obs_locs <- spatstat::superimpose(presence, absence, check = FALSE)
  spatstat::marks(obs_locs)$id <- seq(1, obs_locs$n, 1)
  spatstat::marks(obs_locs) <- spatstat::marks(obs_locs)[ , c(4, 2, 3, 1)]
 
# Specify categories for varying degrees of spatial uncertainty
## Creates three groups
  spatstat::marks(obs_locs)$levels <- as.factor(stats::rpois(obs_locs$n,
                                                                  lambda = 0.05))
                                                                  
# Run perlrren
  test_perlrren <- perlrren(obs_ppp = obs_locs,
                            covariates = ims,
                            radii = c(10, 100, 500),
                            n_sim = 10)
}

}
