% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv.R
\name{plot_cv}
\alias{plot_cv}
\title{Visualizations for the prediction diagnostics of an estimated ecological niche}
\usage{
plot_cv(input, alpha = 0.05)
}
\arguments{
\item{input}{An object of class 'list' from the \code{\link{lrren}} function.}

\item{alpha}{Numeric. The two-tailed alpha level for significance threshold (default is 0.05).}
}
\value{
This function produces two plots: 1) area under the receiver operating characteristic curve, and 2) precision-recall curve. Each plot shows predictions for the log relative risk surface. The red-colored lines are the average curves.
}
\description{
Create multiple plots of output from the \code{\link{lrren}} function, specifically for the internal k-fold cross-validation diagnostics.
}
\examples{
if (interactive()) {
  plot_cv(input = test_lrren)
}

}
