% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perturb.R
\name{plot_perturb}
\alias{plot_perturb}
\title{Visualizations for a simulated ecological niche after iteratively perturbing the observation coordinates}
\usage{
plot_perturb(
  input,
  predict = TRUE,
  mean_cols = c("#8B3A3A", "#CCCCCC", "#0000CD"),
  var_cols = c("#E5E5E5", "#1A1A1A"),
  cov_labs = c("V1", "V2"),
  cref0 = "+init=epsg:4326",
  cref1 = NULL,
  digits = 1,
  ...
)
}
\arguments{
\item{input}{An object of class 'list' from the \code{\link{perlrren}} function.}

\item{predict}{Logical. If TRUE (the default), will visualize the four summary statistics in geographic space. If FALSE, will not.}

\item{mean_cols}{Character string of length three (3) specifying the colors for plots with a divergent color palette: 1) presence, 2) neither, and 3) absence. The default colors in hex are \code{c("#8B3A3A", "#CCCCCC", "#0000CD")} or \code{c("indianred4", "grey80", "blue3")}.}

\item{var_cols}{Character string of length two (2) specifying the colors for plots with a sequential color palette from low to high values. The default colors in hex are \code{c("#E5E5E5", "#1A1A1A")} or \code{c("grey90", "grey10")}.}

\item{cov_labs}{Character string of length two (2) specifying the x- and y-axis labels in plots of the ecological niche in covariate space. The default values are generic \code{c("V1", "V2")}.}

\item{cref0}{Character. The Coordinate Reference System (CRS) for the x- and y-coordinates in geographic space. The default is WGS84 \code{"+init=epsg:4326"}.}

\item{cref1}{Optional, character. The Coordinate Reference System (CRS) to spatially project the x- and y-coordinates in geographic space.}

\item{digits}{Integer. The number of significant digits for the color key labels using the \code{\link[base]{round}} function (default is 1).}

\item{...}{Arguments passed to \code{\link[fields]{image.plot}} for additional graphical features.}
}
\value{
This function produces four plots in a two-dimensional space where the axes are the two specified covariates: 1) mean of the log relative risk, 2) standard deviation of the log relative risk, 3) mean of the asymptotically normal p-value, and 4) proportion of iterations were statistically significant based on a two-tailed alpha-level threshold. If \code{predict = TRUE}, this function produces an additional four plots of the summary statistics above in a two-dimensional geographic space where the axes are longitude and latitude.
}
\description{
Create multiple plots of output from the \code{\link{perlrren}} function, specifically for the four summary statistics in covariate space and geographic space.
}
\examples{
\donttest{
  plot_perturb(input = test_perlrren)
}

}
