% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.r
\name{purge_envi}
\alias{purge_envi}
\title{Remove envi Configuration and Environments}
\usage{
purge_envi(confirm = TRUE, remove_internal_vars = TRUE)
}
\arguments{
\item{confirm}{should the user be prompted to make sure they want to purge
all envi environments? (Default TRUE)}

\item{remove_internal_vars}{should the envi-internal environment be removed?
(Default TRUE)}
}
\description{
Remove envi Configuration and Environments
}
