% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs.R
\name{plot_obs}
\alias{plot_obs}
\title{Visualizations for an estimated ecological niche in covariate space}
\usage{
plot_obs(
  input,
  plot_cols = c("#8b3a3a", "#cccccc", "#0000cd"),
  alpha = 0.05,
  ...
)
}
\arguments{
\item{input}{An object of class "list" from the \code{\link{lrren}} function.}

\item{plot_cols}{Character string of length three (3) specifying the colors for plotting: 1) presense, 2) neither, and 3) absence. The default colors in hex are \code{c("#8b3a3a", "#cccccc", "#0000cd")} or \code{c("indianred4", "grey80", "blue3")}.}

\item{alpha}{Numeric. The two-tailed alpha level for significance threshold (default is 0.05).}

\item{...}{Arguments passed to \code{\link[spatstat.core]{plot.ppp}} and \code{\link[fields]{image.plot}} for additional graphical features.}
}
\value{
This function produces three plots in a two-dimensional space where the axes are the two specified covariates: 1) observation locations by group, 2) log relative risk surface, and 3) significant p-value surface.
}
\description{
Create multiple plots of output from the \code{\link{lrren}} function, specifically for the observation data and estimated ecological niche.
}
\examples{
\donttest{
plot_obs(input = test_lrren)
}

}
