% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.r
\name{purge_envi}
\alias{purge_envi}
\title{Remove envi Configuration and Environments}
\usage{
purge_envi(confirm = TRUE)
}
\arguments{
\item{confirm}{should the user be prompted to make sure they want to purge
all envi environments? (Default TRUE)}
}
\description{
Remove envi Configuration and Environments
}
